completed Module 7 Lab 2: Custom Hook
