+++ labs/src/hooks/useData.jsx
@@ -1,0 +1,25 @@
+import { useState, useEffect } from "react";
+
+export function useData(url) {
+    const [data, setData] = useState(null)
+
+    useEffect(() => {
+        if(url) {
+            let ignore = false;
+            fetch(url)
+                .then(response => response.json())
+                .then(json => {
+                    if(!ignore) {
+                        setData(json)
+                    }
+                })
+            
+            return () => {
+                ignore = true
+                console.log('Cleanup effect')
+            }
+        }
+    }, [url])
+
+    return data
+}
